imageResize/requirements.txt
Pillow==2.2.1

## pip install -r requirements.txt